+++ lib/datamappify/data/criteria/relational/save_by_key.rb
@@ -1,12 +1,14 @@
+require 'datamappify/data/criteria/relational/concerns/set_criteria'
+
 module Datamappify
   module Data
     module Criteria
       module Relational
         module SaveByKey
-          def initialize(source_class, entity, attributes, &block)
-            super(source_class, entity, {}, attributes, &block)
+          include Concerns::SetCriteria
 
-            @criteria = { key_name => entity.id } unless entity.id.nil?
+          def initialize(source_class, entity, attributes, options = {}, &block)
+            super(source_class, entity, {}, attributes, options, &block)
           end
         end
       end