+++ lib/datamappify/data/mapper.rb
@@ -78,24 +78,26 @@ module Datamappify
       # @return [Array<Attribute>]
       def default_attributes
         @default_attributes ||= default_attribute_names.collect do |attribute|
-          Attribute.new(attribute, default_source_for(attribute), default_source_class)
+          Attribute.new(attribute, default_source_for(attribute), :primary_source_class => default_source_class)
         end
       end
 
       # @return [Array<Attribute>]
       def custom_attributes
-        @custom_attributes ||= custom_mapping.collect do |attribute, source|
-          map_custom_attribute(attribute, source)
+        @custom_attributes ||= custom_mapping.collect do |attribute, source_and_options|
+          map_custom_attribute(attribute, *source_and_options)
         end
       end
 
       # @param (see Data::Mapper::Attribute#initialize)
       #
       # @return [Attribute]
-      def map_custom_attribute(name, source)
+      def map_custom_attribute(name, source, options)
         @custom_attribute_names << name
 
-        Attribute.new(name, source, default_source_class)
+        options.merge!(:primary_source_class => default_source_class)
+
+        Attribute.new(name, source, options)
       end
 
       # @param attribute [Symbol]