+++ spec/support/entities/author.rb
@@ -1,0 +1,6 @@
+class Author
+  include Datamappify::Entity
+
+  attribute :name, String
+  attribute :bio,  String
+end