+++ spec/support/entities/post.rb
@@ -1,0 +1,7 @@
+class Post
+  include Datamappify::Entity
+
+  attribute :title,       String
+
+  attributes_from Author, :prefix_with => :author
+end