+++ spec/support/repositories/active_record/author_repository.rb
@@ -1,0 +1,6 @@
+class AuthorRepositoryActiveRecord
+  include Datamappify::Repository
+
+  for_entity Author
+  default_provider :ActiveRecord
+end