+++ spec/support/repositories/sequel/author_repository.rb
@@ -1,0 +1,6 @@
+class AuthorRepositorySequel
+  include Datamappify::Repository
+
+  for_entity Author
+  default_provider :Sequel
+end